requirements.txt
requests==2.26.0
Pillow==8.2.0
